:art: 많은 요소 수정
console.log 빼고, CSS 수정 작업

+++ src/components/feedsComment/write/FeedsCommentWrite.queries.ts
@@ -19,7 +19,7 @@ export const Q_FETCH_USER = gql`
   }
 `;
 export const Q_FETCH_COMMENTS = gql`
-  query fetchComments($page: Float, $feedId: String!) {
+  query fetchComments($page: Int, $feedId: String!) {
     fetchComments(page: $page, feedId: $feedId) {
       comments {
         id
